Changed date time picker to allow it to contain invalid dates.
The code that uses the date time picker is responsible for handling errors
raised when invalid dates are entered.

The date time picker behaves much more predictable after this change. Before it
tried to set dates to valid ones when an invalid was entered and sometimes the
date you got back with GetValue was not the same as the one shown in the
control.